+++ shortcircuits-converters/src/test/java/com/powsybl/shortcircuits/converters/CgmesImporterTest.java
@@ -4,7 +4,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.shortcircuits.importers;
+package com.powsybl.shortcircuits.converters;
 
 import com.powsybl.cgmes.conformity.test.Cgmes3Catalog;
 import com.powsybl.cgmes.conformity.test.CgmesConformity1Catalog;
